refactoring and making clustering work
